Allow flight recorder to call elasped_time again.
Previously we had to avoid elapsed_time because it would cause
the dump to get stuck inside cuda if the dump was initiated when
cuda was stuck waiting on nccl communication. However, elapsed_time
only got stuck because we were not setting the device correctly before
calling cudaEventElapsedTime. Now that this is fixed inside of elasped_time,
it no longer gets stucks. This means we can simplify handling of duration
timing in the flight recorder.

ghstack-source-id: 8f3afa450e3f6134091775d3bd245bbc0eba0aea
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122731