Add torch.library.custom_op
This is the entrypoint for defining an opaque/blackbox (e.g. PyTorch will
never peek into it) custom op. In this PR, you can specify backend impls
and the abstract impl for this op.

NB: most of this PR is docstrings, please don't be intimidated by the
line count.

There are a number of interesting features:
- we infer the schema from type hints. In a followup I add the ability
  to manually specify a schema.
- we infer the qualname for the operator from the provided function. In a
  followup, we'll add the ability to manually specify a qualname.
- custom_op registrations can override each other. This makes them
  more pleasant to work with in environments like colab.
- we require that the outputs of the custom_op do not alias any inputs
  or each other. We enforce this via a runtime check, but can relax this
  into an opcheck test if it really matters in the future.

Test Plan:
- new tests

ghstack-source-id: d1a3cccfcaf60b2dd33cdc216513d52e6e3a550f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122344